update react types version to v18
The first version of react-query that is compatible with the new major
version of react types is 3.35.0. I updated all the dev dependencies
accordingly to fix the tests. Technically the peer dependencies should
not be affacted but just to be safe I updated them as well so that the
dev dependencies aren't higher than the peer dependencies.