tRPC explained from a REST API & Who is this for
new sections:
- tRPC explained from a REST API 
Please check the section on mistakes ( I am not a native English speaker) and check if you agree with the content  

- who is this for? 
I copied who is this for from further reading. I think the whole further reading page should be moved to the introduction. 
I think the 'relationship with grahpql' in further reading should maybe combined with tRPC explained from a GraphQL background'   

- the ending is a little abrupt of this page