[www] features expanded
> "prepare for trouble, make it double " (+1 if you know the reference) 

So a week ago I noticed that the features section was consisting of 3 features. I am convinced that tRPC has more that 3 features (advantages). What I really noticed was that the mentioned features contained multiple features in it self. For example auto type safety also contained auto completion. Light & snappy are two features. Batteries included was also merged with Framework agnostic. So this pull requests consist of 6 features spreading them out to go deeper in on the features of tRPC.  

Note: Please change it to your liking, as you guys are far more aware of the features of tRPC and which features you as tRPC want to highlight 
Note: Please check grammar ect. 

I no I did not make a  pre announcement that I would edit this, so do with this pull request whatever you want