change tanstack query key to array
As per https://github.com/trpc/trpc/issues/2611 this commit changes the
Path element of the query key used in Tanstack Query to be an Array.

This allows us to take action on related queries such as invalidating
all queries on a router.

In this commit we have implemented this at the internal hooks level so
it reflects across both the `proxy` and `legacy` implementation without
any breaking changes to those api's. Implementing it this way also
allows us to keep code shared between both implementations which we may
not be able to do if implemented further down the chain.

The only breaking change from this would be if the someone is delving in
and using the query keys of Tanstack Query not through the tRPC provided
API's. In this case users will see keys change from something like
`["post.list"]` to `[["post","list"]]