+ all possible input types for router.invalidate()
This was an absolute beast of a typescript types task even though it
looks like a relatively small amount of code 😫. The only way to get the
types for all queries `router.invalidate()` may operate on is to
recursively traverse the router type digging in to any sub routers
(unknown depth) inferring all the input types as you go. Once done you
have another heavily nested type that needs to be flattened so it can be
used as the type for the input which is not easily done with TS Types.
With a load of research the way in the commit has been generated 👍.
Lots of learning on my part ...

This method may be useful in the future for other operations such as
`setQueriesData` which will need to infer output types (Although I feel
this should really get the current query key being updated passed too
that we could use in a discriminated union to know what data type we are
operating on right now! ... 🤷 .. one for the future... Tweeted TKDodo!)

Either way - End result is you get good types for the possible things
you can filter inputs on as well as handling repeated property names
with different types! 👍