change InferAllRouterQueryInputTypes 2 avoid never
As per the style now used in the rest of tRPC introduced by @sachinraja
's Turbo Types we now use filter and avoid never.

Only thing I'm a little sad about is if you do an `infer QueryOrRouter`
Typescript can't seem to then infer that if it's not a query it is a
router and you have to manually do `extends router` which then
introduces a never 👎. To avoid this I have had to keep using
`TRouter['_def']['record'][TKey]` which is not as clear for a developer
reading it.