WIP | start of big query vs infinite query changes
In this commit we have started to look at what impact adding the query
type into the reactQuery query key would have.

A change has been made to allow for selection of all cache items ('all')
as well as individual queries and infinite queries as things like
invalidate may want to work across all. The array query key has also
been altered to allow this behavior and in a way supported by react
queries fuzzy Matching.

I have also had a go at creating some functions that can work back from
reactQuery array keys to known constituents as we may need to do this
user land ... This is to investigate and know if it is possible and
influence the design of these array keys.

However, for this to be implemented in this initial direction we will
need to change all of the react / next client implementations to also
pass in a `queryType` as well as a path and input as this is the only
point we know if we are trying to target only queries, infinite queries
or all. This is going to be a fairly big change unless I can come up
with a smarter way of doing it so wanted to share early to see if brains
smarter then I have a better idea before I get around to looking again!

p.s sorry for the large single commit ... Plane was landing and didn't
have time to split up the commits into a nice story 😂