docs: Export type `Context` for use in `initTRPC`
In the docs, the type `Context` is declared in `server/context.ts`, without being exported. However it's then imported in the code snippet in `server/routers/_app.ts`, so for consistency and clarity I think the type declaration in `server/context.ts` needs to be exported. This PR does that.