remove use of `eval`
I'm not sure if the suggested change is equivalent to what was documented before. However, I think it's important documentation does not suggest using `eval` due to the inherent security risks associated with the method.

Even if the example was included for simplicity and not as a "best practice", people are prone to copy-paste as a starting point, potentially never fixing the usage to something more robust after the fact.