Fix issues with git tags when publishing a new version
We have a custom script which tries to preserve exact versions between packages/*, but constrained versions for examples/*. That's good – that script worked.

The problem was this script was running on the postversion lifecycle hook, i.e. after lerna had committed the new version to Git and tagged the new version.

There was then some extra Git-fu in the postversion hook to try and correct that, but all it was doing is undoing the tagged commit and making a new commit. But the tag was still left pointing to the undone commit.

This commit fixes that issue by jumping out of the postversion hook and not trying any Git-fu at all. Instead, we just jump straight on the 'version' lifecycle hook: https://github.com/lerna/lerna/tree/main/libs/commands/version#lifecycle-scripts

This 'version' hook runs after Lerna has bumped all the versions, but before it commits anything to Git. Our script runs and sets exact versions and is then included in the initial commit to Git. The tags are then applied to the right commit, and there is no longer an opportunity for tags, branches, and commits to get confused.