Fix: Set display to inline-block for code element
This change sets the `display` property of the `code` element to `inline-block`, which makes the right padding of its child `div` (class="line") elements effective and improves user experience expecially on mobile devices.

Before this change, the right padding of the `div` (class="line") elements inside the `code` element was not effective because the `code` element had a default `display` property of `inline`, which does not allow for width or height. This made some lines of code look like they were cut off especially on mobile devices since the padding-right wasn't effective. By setting the `display` property of the `code` element to `inline-block`, the element now has a width and height, and its child `div` (class="line") elements' padding is effective.

I tested this change locally by adding the `display: inline-block` property to the `code` element in the CSS file and verifying that the right padding of its child `div` (class="line") elements was effective.