Proposal/feat: standaloneMiddleware
Enables creation of middlewares that are not bound to any particular
instance of 't'; instead they can define a type of Context that they
expect, and as long as the preceding middleware chain provides a ctx
type that is assignable to the middleware's required context type,
the middleware can be used in the chain.

This commit provides a few tests that demonstrate the usage of
standalone middlewares in two different tRPC instances.