Feat(server): narrow the type of 'meta' when runtime meta value injected
Currently tRPC middlewares always assume that the type of `opts.meta`
is `Meta | undefined` even when a procedure injects metadata using
`.meta()` or `defaultMeta` is provided to `initTRPC().create()`.

This commit modifies the behavior so that whenever either `procedure.meta()`
or `defaultMeta` is provided, the type of Meta is narrowed in downstream
handlers.

The `RootConfig` created in `initTRPC` is now changed
from `Meta` to `Partial<Meta>`, since this facilitates removing the type:

`_meta: TConfig['_meta'] | undefined`

from the definition of `MiddlewareFunction` (this is needed so that middlewares
aren't always stuck assuming `opts.meta` is a union of `Meta` and `undefined`.

Note: this would have the potential to break users' typechecks if they are doing sth
like:

`if (opts.meta && !opts.meta.foo) { throw new Error("bang!"); }`

because with the changed types, `opts.meta` is Partial<Meta>. To combat this
I changed the definition of `MiddlewareFunction` to `Meta | undefined` ONLY when `meta`
has not been injected by a procedure or by `defaultMeta`, which should preserve
backwards compatibility, as without runtime injection, the runtime value of Meta
will be `undefined`.