Server/experimental: allow a mix of text and files with fileUploadHandler
Sending a FormData object with a mix of files and text fields was not possible
when using `experimental_createFileUploadHandler`, because the handling of
all fields was baked into the `UploadHandler` business logic, and logic for
handling non-file fields was only implemented for the memory upload handler.
Ergo, the file upload handler just ignored all fields that were not files.

This PR does a few things:

1. Moves the logic for handling non-file fields to `parseMultipartFormData`,
   and only file fields are passed to the upload handler. Now non-file fields
   are handled identically regardless of upload handler type.
2. Like with other body parsers, adds `maxBodySize` parameter to the function
   `parseMultipartFormData` that defaults to Infinity. Throws when combined
   size of all fields and files exceeds the maximum and unlinks any files
   that were already uploaded.
3. Exposes `NodeOnDiskFile` as `experimental_NodeOnDiskFile` because `zfd.file()`
   schema from `zod-form-data` throws when `NodeOnDiskFile` is returned.
4. Exposes `MaxBodySizeExceededError`