fix(`react-query`): omit cursor in infinite query key
Infinite query keys are expected to be stable. Including the 'cursor'
into the query key of an infinite query will cause there to be redundant
existing infinite queries for a single useInfiniteQuery() call.

This PR omits the 'cursor', if it exists, for the query keys generated
for an infinite query.