feat: add an onError option to RouterCaller
An error handler can now be provided to both the createCallerFactory and createCaller function, allowing unwrapped errors to be thrown when required, or to respond to errors in some other fashion.
It is passed the same arguments as an ErrorFormatter would be, excluding shape, which wouldn't make sense in this context.