chore: :label: export `TRPCBuilder` class from `@trpc/server` package
I don't know what the reasoning is specfically for keeping the entire
`TRPCBuilder` class private, but given that exposing instances of that class
to the user is required in order to use it, keeping the class private seems
like it might have been to coerce users into using the `initTRPC....` functions
instead of instantiating the class directly.

For this reason, I exported only the type of the class so that the type could
be used to explicitly define context and/or meta types in JavaScript land...

```js
import { initTRPC } from '@trpc/server';

/** @type {import("@trpc/server").TRPCBuilder<{ ctx: MyContext }>} */
const init = initTRPC.context();

const t = init.create();
// Define procedures on `t` like usual with a context of type
// `MyContext`...
````

Since the main purpose of exporting the `TRPCBuilder` class is to allow JS
users to specify the builder type returned from `initTRPC` in their code, it
seemed reasonable to colocate the type with the `initTRPC` singleton.

If exporting the whole class would be preferrable for some reason, that seems
entirely feasible and does not expose any new API surface area since the class
has no constructor.