fix(server): ensure order of observable subscription errors
The `ReadableStreamDefaultController.error()` causes **any** future
interactions with the associated stream to error.
Thus, if the underlying observable receives an error, it would have
caused any `.next()` call on the iterator to throw instantly. In case
of multiple emits, this would cause the error to propagate out of
order.

In addition, this would have caused the `stopped` package in the ws
adapter not to be sent. `await iterator.return?.()` would have thrown.