Update error-handling docs to explain how to change errors. (#6781)
## 🎯 Changes

There was a bug in the error handling docs which indicated you could
change an error in the adapter `onError()` method. This does not seem to
be the case. Instead you need to use error formatting.

<!--
Note: once you create a Pull request, we will automatically fix
auto-fixable lint issues in your branch
-->

## ✅ Checklist

- [ ] I have followed the steps listed in the [Contributing
guide](https://github.com/trpc/trpc/blob/main/CONTRIBUTING.md).
- [x] If necessary, I have added documentation related to the changes
made.
- [ ] I have added or updated the tests related to the changes made.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Clarified the explanation of error handling, specifying that the
`onError` method is for handling—not modifying—errors, and added a
reference to the section on error formatting for further guidance.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->